+++ NEWS.md
@@ -2,8 +2,11 @@
 
 Changes in this minor release:
 
-* 
+* Update to sea object output to add event years.
 
+* Added function percent_scarred for calculating timeseries of number trees recording, scarred, and the percentage
+
+* Update to sample_depth to account for single-series fhx objects
 
 # burnr v0.2.2